DEL-308:update Openshift CI repo for major and minor release only

+++ cmd/openshiftCIRelease.go
@@ -343,24 +343,28 @@ func init() {
 			if err != nil {
 				handleError(err)
 			}
-			var intlyOperatorRepoDir string
-			if intlyOperatorRepoDir, err = c.DoIntlyOperatorUpdate(cmd.Context()); err != nil {
-				handleError(err)
-			}
-			if intlyOperatorRepoDir != "" {
-				fmt.Println("Remove temporary directory:", intlyOperatorRepoDir)
-				if err = os.RemoveAll(intlyOperatorRepoDir); err != nil {
+			if c.version.IsPreRelease() {
+				fmt.Println("Skipping the update to Openshift CI release repo as the release version is not a major or a minor one")
+			} else {
+				var intlyOperatorRepoDir string
+				if intlyOperatorRepoDir, err = c.DoIntlyOperatorUpdate(cmd.Context()); err != nil {
 					handleError(err)
 				}
-			}
-			var ciReleaseRepoDir string
-			if ciReleaseRepoDir, err = c.DoOpenShiftReleaseUpdate(cmd.Context()); err != nil {
-				handleError(err)
-			}
-			if ciReleaseRepoDir != "" {
-				fmt.Println("Remove temporary directory:", ciReleaseRepoDir)
-				if err = os.RemoveAll(ciReleaseRepoDir); err != nil {
+				if intlyOperatorRepoDir != "" {
+					fmt.Println("Remove temporary directory:", intlyOperatorRepoDir)
+					if err = os.RemoveAll(intlyOperatorRepoDir); err != nil {
+						handleError(err)
+					}
+				}
+				var ciReleaseRepoDir string
+				if ciReleaseRepoDir, err = c.DoOpenShiftReleaseUpdate(cmd.Context()); err != nil {
 					handleError(err)
+				}
+				if ciReleaseRepoDir != "" {
+					fmt.Println("Remove temporary directory:", ciReleaseRepoDir)
+					if err = os.RemoveAll(ciReleaseRepoDir); err != nil {
+						handleError(err)
+					}
 				}
 			}
 		},
